move loading of scrolling position to base class
